refactor(home): move all types and interfaces to types.ts
Centralise the domain types, component props and hook param/result types
in types.ts instead of declaring them inline across the files.
